chore(wms): move three-books tailgate services to shared
